+++ aimlflow/utils.py
@@ -81,21 +81,21 @@ def get_mlflow_experiments(client, experiment):
     return experiments
 
 
-def get_aim_run(repo_inst, run_id, experiment_id, run_cache):
+def get_aim_run(repo_inst, run_id, experiment_name, run_cache):
     if run_cache.get(run_id):
         aim_run = Run(
             run_hash=run_cache[run_id],
             repo=repo_inst,
             system_tracking_interval=None,
             capture_terminal_logs=False,
-            experiment=experiment_id,
+            experiment=experiment_name,
         )
     else:
         aim_run = Run(
             repo=repo_inst,
             system_tracking_interval=None,
             capture_terminal_logs=False,
-            experiment=experiment_id,
+            experiment=experiment_name,
         )
         run_cache[run_id] = aim_run.hash
 
@@ -207,7 +207,7 @@ def convert_existing_logs(repo_inst, tracking_uri, experiment=None, no_cache=Fal
         for run in tqdm(runs, desc=f'Parsing mlflow runs for experiment `{ex.name}`', total=len(runs)):
             run_id = run.info.run_id
             # get corresponding `aim.Run` object for mlflow run
-            aim_run = get_aim_run(repo_inst, run_id, ex.experiment_id, run_cache)
+            aim_run = get_aim_run(repo_inst, run_id, ex.name, run_cache)
             # Collect params and tags
             collect_run_params(aim_run, run)