+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+# Unreleased
+- Set mlflow experiment name as `aim.Run`'s experiment (mihran113)
+
 ## 0.1.1 
 - Fix MLFlow active runs' query (mihran113)
 - Add `PyPI` long description (mihran113)
 - Rename `convert` command to `sync` (mihran113)
+
 ## 0.1.0 
 
 - Add mlflow to aim convert command (mihran113)