+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Unreleased
 - Set mlflow experiment name as `aim.Run`'s experiment and parse the mlflow run name (mihran113)
+- Remove 'mlflow_run_name' and add 'mlflow_experiment_id' fields (sirykd)
 
 ## 0.1.1 
 - Fix MLFlow active runs' query (mihran113)