+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Set mlflow experiment name as `aim.Run`'s experiment and parse the mlflow run name (mihran113)
 - Try to parse "string-ified" params values (sirykd)
+- Add "--exclude-artifacts" option to 
 
 ## 0.1.1 
 - Fix MLFlow active runs' query (mihran113)