+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
 ## Unreleased
+- Add "--exclude-artifacts" option to sync command (mihran113)
 - Set mlflow experiment name as `aim.Run`'s experiment and parse the mlflow run name (mihran113)
 - Try to parse "string-ified" params values (sirykd)
-- Add "--exclude-artifacts" option to 
 
 ## 0.1.1 
 - Fix MLFlow active runs' query (mihran113)